Implemented auto-updating
Used electron-updater to prompt the user to install a new update when
the program opens if there is a new GitHub release available with a
higher version number than the current one. Added a node script,
`draft-release`, to automatically draft a release based on the current
source code, automatically packaging a Windows installer.

Problem: when installing for all users, the program installs to Program
Files by default, which requires elevated permissions to change. When
installing an update, the program will close, uninstall the previous
version, and then silently stop. This is addressed in several issues in
the electron-builder repository, but it seems like a perfect fix doesn't
exist yet. Unfortunately, we have to ask the user for administrator
permissions on every run to avoid this problem.